+++ lib/core/src/runtime/mod.rs
@@ -4,4 +4,7 @@ mod core;
 mod event_loop;
 mod handlers;
 
+pub mod test;
+
 pub use core::Runtime;
+pub use test::{TestResult, TestRuntime, TestRuntimeBuilder};